Ensuring symmetry of inline vs. attached files
This commit verifies that a file uploaded via senior_thesis#new and
a file uploaded via generic_files#new will display similarly in the
senior_thesis#show related files section (i.e. filename is the name
of the originally uploaded file; title was what was specified in the
form)

Prior to the commit, this was not the case.