Attempting to apply parity of code.
The FileContentDatastream was implemented in such a way as to assume
that super was in fact going to work. I would prefer that to not be the
case. After all, super was working because FileContentDatastream was
composed of modules that extended/included ActiveSupport::Concern.

What I like about the new FileContentDatastream#run_fits! method is that
the contents of the method are operating at the same level of
abstraction.

What I don't like is that it is perhaps more opaque than the traditional
super method.

I would prefer for the two lines in #run_fits! to have the same
signatures.