Finalizing AccessRights for persisted vs. un-per
In its current state, the access rights reflect the negotiation of
a permissionable object's edit form. As such the AccessRight sits
between the persistence concern and the access concern.

All of this was done to clean up a terribly ugly view and the sprinkling
of hash driven permission checking. I'm working towards creating a
canonical location for the AccessRights of an object