Replacing Doi #save with #save!
The return value of the save was not being checked; this could result
in an unexpected state (i.e. a DOI was registered with the agent,
but the DOI was not persisted in our database; and we wouldn't know
about it. At least with an exception, we have a chance of figuring out
the corrupted data state).