Adding DB col to dummy User#encrypted_password
By not persisting the encrypted_password the DeviseTestHelper#sign_in
was not working. Warden was attempting to serialize the User by ID and
verifying that the substring of the encrypted_password in the session
matched the same substring of the encrypted_password on the User. Since
the previous method was not persisted, the encrypted_password changed
from second to second.