Updating embargo release date for non-embargo
Prior to this patch, it would be possible to specify an embargo release
date but a non-embargo access right. If this was the case, the embargo
release date would persist and could then "show up" again if the object
were transitioned from Private to Open Access. Instead of showing up
as Open Access, the object that had just transitioned would be Open
Access with Embargo Release Date.

IDR-193 #resolve