Adding registry for curation concerns
Instead of relying on updating a constant, Curate now provides a
configuration option for registering new curation_concerns.

This also involved adjusting the `curate` and `curate:work`
generators. The registration occurs in an initializer.

A refactoring was also done to eliminate the duplicate behavior
regarding factory names and the shared examples.