Cauterizing failing tests
There are a handful of failing tests that need to be addressed. One is
related to the alpha ActiveFedora 7. Another is related to a change
in Blacklight/Kaminari. And third, we have a preferred email entry
that is naive in its implementation (i.e. preferred email should simply
be email and removed from the form as it is confusing)