Remove hard-coding view logic on quick classify
Given that Curate allows for customizing which CurationConcerns are
available, we need to be able to handle this customization.

This is a step towards that. We are envisioning that the
.quick-classify area could vary based on the context (i.e. Professor
Pie only deposits Research Paper and Datasets so lets have that be the
default; Whereas Apple is a new user and sees a different set of
options for classification)

It should be noted that prior to this change CurateND was broke
because it did not register an image as a valid work type.

Adding a view spec to validate render behavior.