Expanding Rails 4.0.x support
Rails ~>4.0.1 is enforcing parameters when using a member path method
(i.e. person_path(user.person) where user.person was not persisted)

By switching to using the User#repository_noid we also do not need to
load the Person from Fedora on each request (as that is part of the
action bar).

closes #303
closes #304