Add soft delete option to application template
Edited the application template to prompt the user to enable soft
delete functionality.  The user will only be asked to enable delete
if they chose to install and unzip jetty.  This is because jetty
(Fedora) needs to be in place so we have a place to copy the xacml
files.

The monkey patch and xacml files are stored in the
./lib/generatator/curate/soft_delete directory so the application
template can download the files from the online repository.

HYDRASIR-122 #close