Resolves Rubydora reload and API problem with deletes (soft and hard).
With the upgrade of Rubydora from 1.6.5 to 1.7.0 they switched to an
API approach to method calls.  We had to adjust the monkey patch to
account for this.

They removed the automatic reload after actions in Fedora.  We had
to adjust the with_generic_file controller to use a before action.

HYDRASIR-237 #close