Add representative image to Collection objects
Add create and update methods to the CollectionController to call
add_profile_image if an image is uploaded. Also add an index method to
redirect to the "My Collections" catalog view in order to display
Collection thumbnails (rather than modify the Blacklight method).
Include Hydra::Derivatives with the Collection model and add the
necessary datastreams, accessors and methods; used Person model as a
template. Add upload field to the Collection form. Add the
thumbnail display to the Collection show view and add another check
to the thumbnail partial (check for the presence of the 'thumbnail'
datastream). Add a new test to the collection feature spec.