HYDRASIR-394 Profile sections should not have thumbnails/avatar
Modified new/edit collection form to not show avatar controls when the collection is a ProfileSection.
Modified collections_controller create method to skip the setting of the .file when the collection is really a ProfileSection
Modified the show form for collection to not show the avatar for a ProfileSection.
Added spec test to verify avatar control is not present when creating/editing a profile section.