Not looking up User in view
Given that we already know the user we are editing, instead refer
to the object that is being edited (i.e. f.object is the user).

In the case of editing your own profile, devise's convention is to
have the route be /users/edit. This means that `params[:id]` is nil.

[skip ci]