Adjusting Registration updates for managers
In the case where a new user is attempting to sign up, and that user
is a repository manager, the Registration#update action was failing.
This is because `params[:user][:id]` was nil; The
RegistrationController is typically a singular resource (i.e. a
resource in which a given user has one and only one resource, and thus
there is no params[:id] that is passed).

So I am adding some descriptive methods to say what is happening. If
you are not a repository manager we update your account straight away.
If you are a repository manager, we determine if you are updating
another user's account or if you are updating your own account.

This is a stop gap measure as it is rearranging behavior that was
previously moved from Devise into Curate.

Also adjusting how the manager is looked up; By favoring Hash#fetch
over Hash#[], an exception is raised earlier rather than later.