Adjusting registration controller
In the case of a user editing their own account, we need to ask them
to reauthenticate (as the password field is being rendered?).

In the case where I am editing another user's account (i.e. repository
manager), then I am not asking for the password of the account being
edited.