Adding force profile update to catalog
We encountered a halted state when a user would signin but not yet
profile their full profile information. What would happen is they
would be redirected, without notification, to the account update page.

If they went to the catalog controller, while logged in, the SOLR
query would be incorrect.

Adding a flash message to let the user know what happened. And adding
enforcement on the catalog controller.