Better fix for moving works from embargo to public
Undo previous changes to app/repository_models/curation_concern/embargoable.rb
and implement a simpler and better fix for users not being able to
move a work from embargoed status to open/public.  The problem is
that the setter method for embargo_release_date= gets called twice
and on the second call it adds the embargo date again to
@embargo_release_date.  So I added a conditional that only sets
@embargo_release_date once (to nil) if the user is trying to take
something out of embargo.

Fixing embargoable.rb caused some of the embargo tests to break even
though the app was performing as it should.  So I tweaked the tests
accordingly.