Adding multi-value presence validation
Given that we are allowing multi-value for a given attribute, we
need to better handle the presence of that attribute.

Consider:

    class MyModel
      # Assume this class supports ActiveModel validations
      validate :contributor, presence: true
    end

    my_model = MyModel.new
    my_model.contributor = ['']
    my_model.valid?
    => true

When in fact we likely mean for it not to be true.

Solution:
    class MyModel
      # Assume this class supports ActiveModel validations
      validate :contributor, multi_value_presence: true
    end

    my_model = MyModel.new
    my_model.contributor = ['']
    my_model.valid?
    => false

[skip ci]
This commit was an amended commit to include a better commit message.
The build that verifies the actual code changes can be found at:
https://travis-ci.org/projecthydra-labs/curate/builds/25186700