Make button show text optional in the partial.
The text "add to collections" should only show on the individual
concern's show page button. The text should not show up on the
root page, only the icon should be displayed to be consistent with
the style of the resto of the buttons. To address this, I've added a
local variable 'use_text' that can be passed when calling the partial.
It defaults to false which does not show the text.

dltp-52#