Replacing usage of user.email with user.user_key
Throughout the Curate code, there are cases where User#email is
treated as the user key. This works for instances where devise is
configured to use email as the user identifier.

In the case of CurateND, we are using a login/netid. This means we may
encounter Group or Embargo related errors, among other odd things. If
we are doing something with the email, other than displaying in the
UI, the method likely needs changing.

The configuration for what User#user_key is, is defined in the devise
config of CurateND https://github.com/ndlib/curate_nd/blob/master/config/initializers/devise.rb#L38

```ruby
config.authentication_keys = [ :username ]
```

DLTP-137 #closes