Change Help button Functionality
Change the Help button to a dropdown menu that contains a link to FAQs:
a page that is to be written later, and Request Help which provides the
previous action of the help button, launching a modal help window.

The help button/dropdown menu has also been made available when a user
is not logged in.

dltp-174#
dltp-170#

Button Alignment

Align dropdown for Help! to the right edge of the button so that it does
not overflow off the page. From DBH feedback.

Correct URL on line 8 from collections_path to new_help_request_path.